add weights to mds, relocate pdb functions to utils clea n notebook
